+++ Core.lua
@@ -1,7 +1,13 @@
 -- TODO: Check conventions https://github.com/luarocks/lua-style-guide
 -- TODO: Remove unused Ace3 Libs
+
+-- TODO: Auto combine fragments and tell user if there are more in their bank
+-- TODO: Cleanup options layout
+-- TODO: Update translations
+-- TODO: Destroy utilities frame and recreate to remove /reload requirement
 TheGrimRepair = LibStub("AceAddon-3.0"):NewAddon("TGR", "AceConsole-3.0", "AceEvent-3.0")
 local L = LibStub("AceLocale-3.0"):GetLocale("TGR", true)
+local AceGUI = LibStub("AceGUI-3.0")
 
 local defaults = {
     profile = {
@@ -9,6 +15,9 @@ local defaults = {
         is_selling_gray_items = true,
         is_showing_sale_details = true,
         is_keeping_transmog_items = true,
+        is_showing_with_bags = false,
+        is_showing_with_merchants = true,
+        is_df_combine_fragments = true,
     }
 }
 
@@ -20,7 +29,7 @@ local options = {
         desc = {
             order = 1,
             type = "description",
-            name = L["MSG_TIP_DISABLE"],
+            name = L["MSG_TIP_DISABLE"] .. "\n\n",
         },
         general = {
             type = "group",
@@ -85,10 +94,81 @@ local options = {
                         },
                     },
                 },
+                group_tgr_utilities = {
+                    order = 3,
+                    type = "group",
+                    inline = true,
+                    name = "TheGrimRepair Utilities",
+                    args = {
+                        show_with_bags = {
+                            order = 1,
+                            width = "double",
+                            type = "toggle",
+                            name = "Show with all bag events",
+                            desc = "If enabled, the utilities window shows and hides with all bag events (including merchant visits)",
+                            get = "is_showing_with_bags",
+                            set = "toggle_show_with_bags",
+                        },
+                        show_with_merchants = {
+                            order = 2,
+                            width = "double",
+                            type = "toggle",
+                            name = "Show with merchants",
+                            desc = "If enabled, the utilities window shows and hides with merchant visits",
+                            disabled = "is_showing_with_bags",
+                            get = "is_showing_with_merchants",
+                            set = "toggle_show_with_merchants",
+                        },
+                    },
+                },
+            },
+        },
+        xpac_dragonflight = {
+            type = "group",
+            order = 3,
+            name = "Dragonflight Options",
+            args = {
+                df_group_tgr_utilities = {
+                    order = 1,
+                    type = "group",
+                    inline = true,
+                    name = "TheGrimRepair Utilities",
+                    args = {
+                        df_combine_fragments = {
+                            order = 1,
+                            width = "double",
+                            type = "toggle",
+                            name = "Combine Shadowflame Crest Fragments",
+                            desc = "If enabled, the utilites window shows an option to combine Shadowflame Crest Fragments",
+                            get = "is_df_combine_fragments",
+                            set = "toggle_df_combine_fragments",
+                        },
+                    },
+                },
+                df_reload_tip = {
+                    order = 2,
+                    width = "full",
+                    type = "description",
+                    name = "Tip: If you made any changes here, you will need to /reload to see them\n\n",
+                },
+                df_development_header = {
+                    order = 3,
+                    width = "full",
+                    type = "header",
+                    name = "THIS FEATURE IS UNDER ACTIVE DEVELOPMENT",
+                },
+                df_development_text = {
+                    order = 4,
+                    width = "full",
+                    type = "description",
+                    name = "In the future, I would like to make the TheGrimRepair Utilities window smarter than just fancy macros.\n\nBlizzard requires user interaction in some circumstances, some functionality will require you to interact more than once.",
+                },
             },
         },
     },
 }
+
+local tgr_frame
 
 function TheGrimRepair:OnInitialize()
     self.db = LibStub("AceDB-3.0"):New("TheGrimRepairDB", defaults, true)
@@ -101,10 +181,33 @@ function TheGrimRepair:OnInitialize()
 
     self:RegisterChatCommand("tgr", "slash_command")
     self:RegisterChatCommand("thegrimrepair", "slash_command")
+    self:RegisterChatCommand("tgru", "show_utilities")
 end
 
 function TheGrimRepair:OnEnable()
     self:RegisterEvent("PLAYER_INTERACTION_MANAGER_FRAME_SHOW")
+    self:RegisterEvent("MERCHANT_CLOSED")
+
+    EventRegistry:RegisterCallback("ContainerFrame.OpenAllBags", function()
+        self:open_utilities()
+    end)
+    EventRegistry:RegisterCallback("ContainerFrame.CloseAllBags", function()
+        self:close_utilities()
+    end)
+end
+
+function TheGrimRepair:open_utilities()
+    local is_showing_with_bags = self:is_showing_with_bags()
+    if is_showing_with_bags then
+        self:show_utilities()
+    end
+end
+
+function TheGrimRepair:close_utilities()
+    local is_showing_with_bags = self:is_showing_with_bags()
+    if is_showing_with_bags then
+        self:hide_utilities()
+    end
 end
 
 function TheGrimRepair:OnDisable()
@@ -112,9 +215,11 @@ function TheGrimRepair:OnDisable()
 end
 
 function TheGrimRepair:PLAYER_INTERACTION_MANAGER_FRAME_SHOW(event, typeId)
+    local is_showing_with_merchants = self:is_showing_with_merchants()
+
     -- Hold shift to disable the addon during this interaction
     if IsShiftKeyDown() then
-        self:Print(L["MSG_SKIPPED"])
+        self:Print(L["MSG_SKIPPED"]) --TODO: TheGrimRepair was skipped
         return
     end
 
@@ -122,11 +227,42 @@ function TheGrimRepair:PLAYER_INTERACTION_MANAGER_FRAME_SHOW(event, typeId)
     if typeId == Enum.PlayerInteractionType.Merchant then
         self:auto_repair()
         self:auto_sell()
+        if is_showing_with_merchants then
+            self:show_utilities()
+        end
+    end
+end
+
+function TheGrimRepair:MERCHANT_CLOSED(event)
+    local is_showing_with_merchants = self:is_showing_with_merchants()
+
+    if is_showing_with_merchants then
+        self:hide_utilities()
     end
 end
 
 function TheGrimRepair_OnAddonCompartmentClick(addonName, buttonName)
-    InterfaceOptionsFrame_OpenToCategory("TheGrimRepair")
+    if buttonName == "RightButton" then
+        TheGrimRepair:show_utilities()
+    else
+        InterfaceOptionsFrame_OpenToCategory("TheGrimRepair")
+    end
+end
+
+function TheGrimRepair_OnAddonCompartmentEnter(addonName, buttonName)
+	if not tooltip then
+		tooltip = CreateFrame("GameTooltip", "TheGrimRepair_AddonCompartimentTooltip", UIParent, "GameTooltipTemplate")
+	end
+
+    tooltip:SetOwner(buttonName, "ANCHOR_LEFT");
+	tooltip:SetText("TheGrimRepair")
+	tooltip:AddLine("Left-click opens options", 1, 1, 1)
+	tooltip:AddLine("Right-click opens TheGrimRepair Utilities", 1, 1, 1)
+	tooltip:Show()
+end
+
+function TheGrimRepair_OnAddonCompartmentLeave(addonName, buttonName)
+	tooltip:Hide()
 end
 
 function TheGrimRepair:is_using_guild_repairs(info)
@@ -194,6 +330,30 @@ function TheGrimRepair:toggle_show_sale_details(info, value)
     self.db.profile.is_showing_sale_details = value
 end
 
+function TheGrimRepair:is_showing_with_bags(info)
+    return self.db.profile.is_showing_with_bags
+end
+
+function TheGrimRepair:toggle_show_with_bags(info, value)
+    self.db.profile.is_showing_with_bags = value
+end
+
+function TheGrimRepair:is_showing_with_merchants(info)
+    return self.db.profile.is_showing_with_merchants
+end
+
+function TheGrimRepair:toggle_show_with_merchants(info, value)
+    self.db.profile.is_showing_with_merchants = value
+end
+
+function TheGrimRepair:is_df_combine_fragments(info)
+    return self.db.profile.is_df_combine_fragments
+end
+
+function TheGrimRepair:toggle_df_combine_fragments(info, value)
+    self.db.profile.is_df_combine_fragments = value
+end
+
 function TheGrimRepair:auto_sell()
     local is_selling_gray_items = self:is_selling_gray_items()
     local is_showing_sale_details = self:is_showing_sale_details()
@@ -206,7 +366,7 @@ function TheGrimRepair:auto_sell()
 
     -- Auto Sell Grays
     if is_selling_gray_items then
-        for bag_number = 0, NUM_BAG_SLOTS do
+        for bag_number = 0, NUM_TOTAL_EQUIPPED_BAG_SLOTS do
 
             for slot_number = 1, C_Container.GetContainerNumSlots(bag_number) do
                 local item_location = ItemLocation:CreateFromBagAndSlot(bag_number, slot_number)
@@ -246,6 +406,118 @@ function TheGrimRepair:auto_sell()
     end
 end
 
+function TheGrimRepair:df_find_fragments()
+    local required_fragments = 15
+    local fragment_ids = {
+        204075, --Whelpling's Shadowflame Crest Fragment
+        204076, --Drake's Shadowflame Crest Fragment
+        204077, --Wyrm's Shadowflame Crest Fragment
+        204078, --Aspect's Shadowflame Crest Fragment
+        -- For testing
+        -- 190328, --Rousing Frost
+        -- 190320, --Rousing Fire
+    }
+    local fragment_combine_table = {}
+
+    for bag_number = 0, NUM_TOTAL_EQUIPPED_BAG_SLOTS do
+
+        for slot_number = 1, C_Container.GetContainerNumSlots(bag_number) do
+            local item_location = ItemLocation:CreateFromBagAndSlot(bag_number, slot_number)
+
+            if item_location:IsValid() then
+                local item_id = C_Item.GetItemID(item_location)
+
+                for _, value in pairs(fragment_ids) do
+                    if value == item_id then
+                        local fragment_count = GetItemCount(item_id)
+                        -- local item_name, item_link = GetItemInfo(item_id)
+                        local fragment_combine_count = math.floor(fragment_count / required_fragments)
+
+                        -- Save the item_id with the combine count
+                        table.insert(fragment_combine_table, item_id .. ":" .. fragment_combine_count)
+                        break
+                    end
+                end
+
+            end
+        end
+    end
+    return fragment_combine_table
+end
+
 function TheGrimRepair:slash_command()
     InterfaceOptionsFrame_OpenToCategory(self.optionsFrame)
+end
+
+function TheGrimRepair:show_utilities()
+    if tgr_frame then
+        tgr_frame:Show()
+    else
+        tgr_frame = AceGUI:Create("Frame")
+        tgr_frame:SetTitle("TheGrimRepair Utilities")
+        tgr_frame:SetHeight(200)
+        tgr_frame:SetWidth(400)
+        tgr_frame:SetPoint("TOP", 0, -200)
+        tgr_frame:SetStatusText("Configure in TheGrimRepair options")
+
+        local tgru_text = AceGUI:Create("Label")
+        tgru_text:SetFullWidth(true)
+        tgru_text:SetText("\nTip: Opens with /tgru or by right-clicking from the addon dropdown\n\n")
+        tgr_frame:AddChild(tgru_text)
+
+        local df_heading = AceGUI:Create("Heading")
+        df_heading:SetFullWidth(true)
+        df_heading:SetText("Dragonflight")
+        tgr_frame:AddChild(df_heading)
+
+        local df_combine_fragments_option = self:is_df_combine_fragments()
+
+        if df_combine_fragments_option then
+            local df_fragment_btn = CreateFrame("Button", nil, UIParent, "SecureActionButtonTemplate, UIPanelButtonTemplate")
+            local df_fragment_widget = {
+                frame = df_fragment_btn
+            }
+
+            local df_fragments_text = AceGUI:Create("Label")
+            df_fragments_text:SetFullWidth(true)
+            df_fragments_text:SetText("\nOnly active if you have enough crest fragments to combine:\n\n")
+
+            -- Build /use string dynamically
+            local df_combine_fragments_count = self:df_find_fragments()
+            local df_fragment_use = ""
+
+            -- Parse the response, break out the combine count
+            for _, value in pairs(df_combine_fragments_count) do
+                local item_id, count = value:match("([^:]+):([^:]+)")
+                count = tonumber(count)
+
+                if count > 0 then
+                    df_fragment_use = df_fragment_use .. "\n/use item:" .. item_id
+                end
+            end
+
+            df_fragment_btn:SetAttribute("type", "macro")
+            df_fragment_btn:SetAttribute("macrotext", df_fragment_use)
+            df_fragment_btn:SetText("Combine Shadowflame Crest Fragments")
+            df_fragment_btn:SetSize(366, df_fragment_btn:GetTextHeight() + 9)
+            df_fragment_btn:SetScript("PostClick", function()
+                self:Print("Combining crest fragments... Blizzard requires you to click the button again if you have more to combine.")
+            end)
+
+            AceGUI:RegisterAsWidget(df_fragment_widget)
+            tgr_frame:AddChild(df_fragment_widget)
+        else
+            local df_default_text = AceGUI:Create("Label")
+            df_default_text:SetFullWidth(true)
+            df_default_text:SetText("\nYou haven't configured any Dragonflight options yet or you need to /reload")
+            tgr_frame:AddChild(df_default_text)
+        end
+
+    end
+end
+
+function TheGrimRepair:hide_utilities()
+    if tgr_frame then
+        tgr_frame:Hide()
+    end
 end